Struct tokio_test::io::Mock

pub struct Mock { /* fields omitted */ }

An I/O object that follows a predefined script.

This value is created by Builder and implements AsyncRead + AsyncWrite. It follows the scenario described by the builder and panics otherwise.