Macro tokio_test::assert_ready

macro_rules! assert_ready {
    ($e:expr) => { ... };
    ($e:expr, $($msg:tt)+) => { ... };
}

Asserts a Poll is ready, returning the value.

This will invoke panic! if the provided Poll does not evaluate to Poll::Ready at runtime.

Custom Messages

This macro has a second form, where a custom panic message can be provided with or without arguments for formatting.

Examples

use futures_util::future;
use tokio_test::{assert_ready, task};

let mut fut = task::spawn(future::ready(()));
assert_ready!(fut.poll());