tokio_test::task

Function spawn

pub fn spawn<T>(task: T) -> Spawn<T> 

Spawn a future into a Spawn which wraps the future in a mocked executor.

This can be used to spawn a Future or a Stream.

For more information, check the module docs.