Struct tokio_test::task::Spawn

pub struct Spawn<T> { /* private fields */ }

Future spawned on a mock task

Implementations

impl<T> Spawn<T>

pub fn into_inner(self) -> Twhere
T: Unpin,

Consumes self returning the inner value

pub fn is_woken(&self) -> bool

Returns true if the inner future has received a wake notification since the last call to enter.

pub fn waker_ref_count(&self) -> usize

Returns the number of references to the task waker

The task itself holds a reference. The return value will never be zero.

pub fn enter<F, R>(&mut self, f: F) -> Rwhere
F: FnOnce(&mut Context<'_>, Pin<&mut T>) -> R,

Enter the task context

impl<T: Future> Spawn<T>

pub fn poll(&mut self) -> Poll<T::Output>

Polls a future

impl<T: Stream> Spawn<T>

pub fn poll_next(&mut self) -> Poll<Option<T::Item>>

Polls a stream